dashboard/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
pexpect>=4.8.0
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.0
bcrypt>=4.1.0
PyJWT>=2.8.0
email-validator>=2.0.0
cryptography>=42.0.0
aiobotocore>=2.12.0
python-multipart>=0.0.6
jinja2>=3.1.0
